feature/line/pyproject.toml
[tool.poetry]
name = "flex_message"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
numpy = "^1.26.0"
flask = "^3.1.0"
matplotlib = "^3.9.2"
line-bot-sdk = "^3.14.2"
google-generativeai = "^0.8.3"
langchain-google-genai = "^2.0.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
